fix: Add selenium test file #261
